Upd_070821 * Renamed 'SurfaceMutable' -> 'MutableTextureBuffer'. * Class Bone updated - implemented most of the methods. * Class CutsceneAsset updated - class structure updated with more discovered fields. * Class MutableTextureBuffer updated - implemented 'CalculateMipMaps', 'ReduceColors' methods. * Class PhysSystem updated - implemented 'SetVertexPos' and 'SetNormalPos', more fields discovered. * Class Surface updated - renamed class fields, implemented 'SetPixelColor', 'GetPixelColor' and 'SetSurfaceColorPalette'. * Class TextBox updated - implemented 'SetTextScale' method, documented flags field.

+++ TOD1/Performance.cpp
@@ -46,7 +46,7 @@ INT64 Performance::GetSeconds()
 	}
 
 	if (!MeasuringPerformanceSeconds)
-		return 1000 * timeGetTime();
+		return 1000 * (INT64)timeGetTime();
 
 	LARGE_INTEGER freq;
 	QueryPerformanceCounter(&freq);
